core: Moved Bitmap into its own module for refactoring rendering
